Start element module for rendering.
