+++ ftml/src/render/html/mod.rs
@@ -26,6 +26,7 @@ mod macros;
 
 mod builder;
 mod context;
+mod element;
 mod escape;
 mod meta;
 mod output;
@@ -38,9 +39,10 @@ pub use self::output::HtmlOutput;
 use super::prelude;
 
 use self::context::HtmlContext;
+use self::element::render_element;
 use crate::data::PageInfo;
 use crate::render::Render;
-use crate::tree::{Container, Element, SyntaxTree};
+use crate::tree::SyntaxTree;
 
 #[derive(Debug)]
 pub struct HtmlRender;
@@ -79,32 +81,4 @@ impl Render for HtmlRender {
         // Build and return HtmlOutput
         ctx.into()
     }
-}
-
-fn render_element(ctx: &mut HtmlContext, element: &Element) {
-    match element {
-        Element::Container(container) => render_container(ctx, container),
-        _ => todo!(),
-    }
-}
-
-fn render_container(ctx: &mut HtmlContext, container: &Container) {
-    // Get HTML tag type for this type of container
-    let tag_spec = container.ctype().html_tag();
-
-    // Build the tag
-    let mut tag = ctx.html().tag(tag_spec.tag());
-
-    // Merge the class attribute with the container's class, if it conflicts
-    match tag_spec.class() {
-        Some(class) => tag.attr_map_merge(container.attributes(), ("class", class)),
-        None => tag.attr_map(container.attributes()),
-    };
-
-    // Add container internals
-    tag.contents(|ctx| {
-        for inner_element in container.elements() {
-            render_element(ctx, inner_element);
-        }
-    });
 }
